+++ CopilotKit/packages/react-ui/src/components/chat/Modal.tsx
@@ -63,6 +63,8 @@ export const CopilotModal = ({
   hitEscapeToClose = true,
   onSetOpen,
   onSubmitMessage,
+  onStopGeneration,
+  onReloadMessages,
   shortcut = "/",
   icons,
   labels,
@@ -99,6 +101,8 @@ export const CopilotModal = ({
           <CopilotChat
             instructions={instructions}
             onSubmitMessage={onSubmitMessage}
+            onStopGeneration={onStopGeneration}
+            onReloadMessages={onReloadMessages}
             makeSystemMessage={makeSystemMessage}
             showResponseButton={showResponseButton}
             onInProgress={onInProgress}