+++ CopilotKit/packages/runtime/src/graphql/inputs/message.input.ts
@@ -1,5 +1,5 @@
 import { Field, InputType } from "type-graphql";
-import { MessageRole, ActionExecutionScope } from "../types/enums";
+import { MessageRole } from "../types/enums";
 import { BaseMessageInput } from "../types/base";
 
 // GraphQL does not support union types in inputs, so we need to use
@@ -24,6 +24,9 @@ export class TextMessageInput {
   @Field(() => String)
   content: string;
 
+  @Field(() => String, { nullable: true })
+  parentMessageId?: string;
+
   @Field(() => MessageRole)
   role: MessageRole;
 }
@@ -36,8 +39,14 @@ export class ActionExecutionMessageInput {
   @Field(() => String)
   arguments: string;
 
-  @Field(() => ActionExecutionScope)
-  scope: ActionExecutionScope;
+  @Field(() => String, { nullable: true })
+  parentMessageId?: string;
+
+  @Field(() => String, {
+    nullable: true,
+    deprecationReason: "This field will be removed in a future version",
+  })
+  scope?: String;
 }
 
 @InputType()
@@ -47,6 +56,9 @@ export class ResultMessageInput {
 
   @Field(() => String)
   actionName: string;
+
+  @Field(() => String, { nullable: true })
+  parentMessageId?: string;
 
   @Field(() => String)
   result: string;