+++ examples/coagents-qa-text/agent/greeting_agent/demo.py
@@ -27,4 +27,16 @@ add_fastapi_endpoint(app, sdk, "/copilotkit")
 def main():
     """Run the uvicorn server."""
     port = int(os.getenv("PORT", "8000"))
-    uvicorn.run("greeting_agent.demo:app", host="0.0.0.0", port=port, reload=True)
+    uvicorn.run(
+        "greeting_agent.demo:app",
+        host="0.0.0.0",
+        port=port,
+        reload=True,
+        reload_dirs=(
+            ["."] +
+            (["../../../sdk-python/copilotkit"]
+             if os.path.exists("../../../sdk-python/copilotkit")
+             else []
+             )
+        )
+    )